version boutons products corrigé

+++ src/Controller/AdresseController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 use App\Entity\Adresse;
 use App\Form\AdresseType;
 use App\Repository\AdresseRepository;
+
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -13,6 +14,7 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/adresse')]
 class AdresseController extends AbstractController
 {
+    // cette fonction n'est pas appeler
     #[Route('/adresse_index', name: 'adresse_index', methods: ['GET'])]
     public function index(AdresseRepository $adresseRepository): Response
     {
@@ -22,6 +24,7 @@ class AdresseController extends AbstractController
         $user=$this->getUser();
         return $this->render('adresse/index.html.twig', [
             'adresses' => $adresseRepository->findBy(['user_id' => $user]),
+
         ]);
     }
 
@@ -40,50 +43,88 @@ class AdresseController extends AbstractController
             $entityManager->persist($adresse);
             $entityManager->flush();
 
-            return $this->redirectToRoute('adresse_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('profil', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('adresse/new.html.twig', [
             'adresse' => $adresse,
             'form' => $form,
+
         ]);
     }
 
     #[Route('/adresse_show{id}', name: 'adresse_show', methods: ['GET'])]
     public function show(Adresse $adresse): Response
     {
+        // je dois proteger mon code car j'ai fais des cruds mais je dois les rendrent propres a chaque utilisateurs
+        // je reccupere l'user connecté
+        $userConnected=$this->getUser();
+        // je reccupere l'adresse de l'user connecté
+        $userAdresse=$adresse->getUserId();
+        // et je fais une comparaison :
+        // si l'user connectz est différent de l'id de l'user a qui appartient l'adresse je le logout.
+        if ($userConnected != $userAdresse){
+            return $this->redirectToRoute('logout');
+        }
+        // sinon je lui montre son adresse
         return $this->render('adresse/show.html.twig', [
             'adresse' => $adresse,
+
         ]);
     }
 
     #[Route('/{id}/adresse_edit', name: 'adresse_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Adresse $adresse, EntityManagerInterface $entityManager): Response
     {
+        
+        // je dois proteger mon code car j'ai fais des cruds mais je dois les rendrent propres a chaque utilisateurs
+        // je reccupere l'user connecté
+        $userConnected=$this->getUser();
+        // je reccupere l'adresse de l'user connecté
+        $userAdresse=$adresse->getUserId();
+        // et je fais une comparaison :
+        // si l'user connecté est différent de l'id de l'user a qui appartient l'adresse je le logout.
+        if ($userConnected != $userAdresse){
+            return $this->redirectToRoute('logout');
+        }
+        // sinon je lui permet de modifier son adresse
         $form = $this->createForm(AdresseType::class, $adresse);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
 
-            return $this->redirectToRoute('adresse_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('profil', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('adresse/edit.html.twig', [
             'adresse' => $adresse,
             'form' => $form,
+
         ]);
     }
 
     #[Route('/adresse_delete{id}', name: 'adresse_delete', methods: ['POST'])]
     public function delete(Request $request, Adresse $adresse, EntityManagerInterface $entityManager): Response
     {
+        
+        // je dois proteger mon code car j'ai fais des cruds mais je dois les rendrent propres a chaque utilisateurs
+        // je reccupere l'user connecté
+        $userConnected=$this->getUser();
+        // je reccupere l'adresse de l'user connecté
+        $userAdresse=$adresse->getUserId();
+        // et je fais une comparaison :
+        // si l'user connecté est différent de l'id de l'user a qui appartient l'adresse je le logout.
+        if ($userConnected != $userAdresse){
+            return $this->redirectToRoute('logout');
+        }
+        // sinon je lui permet de supprimer son adresse
         if ($this->isCsrfTokenValid('delete'.$adresse->getId(), $request->request->get('_token'))) {
             $entityManager->remove($adresse);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('adresse_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('profil', [], Response::HTTP_SEE_OTHER);
     }
 
     #[Route('/changer-adresse-delivery', name: 'change_adresse_delivery')]
@@ -98,9 +139,10 @@ class AdresseController extends AbstractController
             // si il yen a pas je lui dis va en creer une
             return $this->redirectToRoute('adresse_new');
         }
-            // si oui je l'envoie vers la liste de ces adresses
-        return $this->render('adresse/index.html.twig', [
+            // si oui je l'envoie vers la vue liste de ces adresses dans le profil
+        return $this->render('profil/account.html.twig', [
             'adresses' => $adresseRepository->findBy(['user_id' => $user]),
+
         ]);
     }
 
@@ -122,8 +164,9 @@ class AdresseController extends AbstractController
         // $select_adresse->setLivraison(true);
         $manager->persist($select_adresse);
         $manager->flush();
-        return $this->render('adresse/index.html.twig', [
+        return $this->render('profil/account.html.twig', [
             'adresses' => $list_adresses,
+ 
         ]);
     }
